Fix: suppress warning
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@756 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CMarkMgr.cpp
@@ -88,7 +88,7 @@ bool CMarkMgr::CheckPrev(void) const
 */
 bool CMarkMgr::CheckNext(void) const
 {
-	for( unsigned int i = m_nCurpos + 1; i < Count(); i++ ){
+	for( int i = m_nCurpos + 1; i < Count(); i++ ){
 		if( m_cMarkChain[ i ].IsValid() )
 			return true;
 	}
@@ -119,7 +119,7 @@ bool CMarkMgr::PrevValid(void)
 */
 bool CMarkMgr::NextValid(void)
 {
-	for( unsigned int i = m_nCurpos + 1; i < Count(); i++ ){
+	for( int i = m_nCurpos + 1; i < Count(); i++ ){
 		if( m_cMarkChain[ i ].IsValid() ){
 			m_nCurpos = i;
 			return true;
@@ -155,7 +155,7 @@ void CMarkMgr::Flush(void)
 void CAutoMarkMgr::Add(const CMark& m)
 {
 	//	Œ»ÝˆÊ’u‚ª“r’†‚ÌŽž
-	if( m_nCurpos < m_cMarkChain.size() ){
+	if( m_nCurpos < (int)m_cMarkChain.size() ){
 		//	Œ»ÝˆÊ’u‚Ü‚Å—v‘f‚ðíœ
 		m_cMarkChain.erase( m_cMarkChain.begin() + m_nCurpos, m_cMarkChain.end() );
 	}
